+++ app/src/main/java/ie/setu/placemark/adapters/PlacemarkAdapter.kt
@@ -1,0 +1,34 @@
+package ie.setu.placemark.adapters
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import ie.setu.placemark.databinding.CardPlacemarkBinding
+import ie.setu.placemark.models.PlacemarkModel
+
+class PlacemarkAdapter constructor(private var placemarks: List<PlacemarkModel>) :
+    RecyclerView.Adapter<PlacemarkAdapter.MainHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
+        val binding = CardPlacemarkBinding
+            .inflate(LayoutInflater.from(parent.context), parent, false)
+
+        return MainHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: MainHolder, position: Int) {
+        val placemark = placemarks[holder.adapterPosition]
+        holder.bind(placemark)
+    }
+
+    override fun getItemCount(): Int = placemarks.size
+
+    class MainHolder(private val binding : CardPlacemarkBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(placemark: PlacemarkModel) {
+            binding.placemarkTitle.text = placemark.title
+            binding.description.text = placemark.description
+        }
+    }
+}
