+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="hint_placemarkTitle">Placemark Title</string>
     <string name="placemarkDescription">Placemark Description</string>
     <string name="menu_addPlacemark">Add</string>
+    <string name="menu_cancelPlacemark">Cancel</string>
 </resources>
